Split range tests and helpers out of extractFunctions.ts
